Juin 2025 (#15)
* Add June 2025 page and update table of contents

Created the 2025_06.md page with new content for June 2025 and updated _toc.yml to set it as the root and include the previous month.

* Update June 2025 newsletter with new content

Added a survey link and request for participation, updated the introduction, and expanded the lists of blog posts, packages, and tutorials with new entries relevant to the Julia community. This update provides more resources and information for readers ahead of the summer break.

* Update 2025_06.md

* Update 2025_06.md

* Update 2025_06.md

* Update 2025_06.md